refactored eatery details by creating separate files for composables
