+++ app/src/main/java/com/cornellappdev/android/eateryblue/ui/components/home/PaymentWidgets.kt
@@ -1,0 +1,59 @@
+package com.cornellappdev.android.eateryblue.ui.components.home
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.ButtonDefaults
+import androidx.compose.material.Icon
+import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.cornellappdev.android.eateryblue.R
+import com.cornellappdev.android.eateryblue.data.models.Eatery
+import com.cornellappdev.android.eateryblue.ui.theme.EateryBlue
+import com.cornellappdev.android.eateryblue.ui.theme.Green
+import com.cornellappdev.android.eateryblue.ui.theme.Red
+
+/**
+ * Payment widgets that are displayed at the top right region eatery details screen
+ */
+@Composable
+fun PaymentWidgets(eatery: Eatery, modifier: Modifier = Modifier, onClick: () -> Unit) {
+    Surface(
+        modifier = modifier.clickable {
+            onClick.invoke()
+        }, shape = CircleShape, color = Color.White
+    ) {
+        Row(
+            modifier = Modifier.padding(horizontal = 12.dp, vertical = 8.dp),
+            horizontalArrangement = Arrangement.spacedBy(ButtonDefaults.IconSpacing)
+        ) {
+            if (eatery.paymentAcceptsMealSwipes == true) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_payment_swipes),
+                    contentDescription = "Accepts Swipes",
+                    tint = EateryBlue
+                )
+            }
+            if (eatery.paymentAcceptsBrbs == true) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_payment_brbs),
+                    contentDescription = "Accepts BRBs",
+                    tint = Red
+                )
+            }
+            if (eatery.paymentAcceptsCash == true) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_payment_cash),
+                    contentDescription = "Accepts Cash",
+                    tint = Green
+                )
+            }
+        }
+    }
+}
